refactor(lib): split request into submodules
